# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PIIP API
  version: 0.1.0
  description: |
    PIIP 사건/증거/리포트/결제 플랫폼 API 초기 스펙.
    - 버전: Draft 0.1
    - 상태: INTERNAL DRAFT (실제 구현과 차이 가능)
    - 인증: Bearer JWT
servers:
  - url: https://api.piip.local
    description: Local / Dev
  - url: https://staging.api.piip.example
    description: Staging
  - url: https://api.piip.example
    description: Production
security:
  - bearerAuth: []
  - oauth2: ["cases:read"]
...
paths:
...
  /reports:
    post:
      tags: [Reports]
      summary: 리포트 초안 생성
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseId, title]
              properties:
                caseId: { type: string, format: uuid }
                title: { type: string }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /reports/{reportId}:
    get:
      tags: [Reports]
      summary: 리포트 조회
      operationId: getReport
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Reports]
      summary: 리포트 블록/본문 수정
      operationId: updateReport
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary: { type: string }
                body: { type: object }
      responses:
        "200":
          description: 수정됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://api.piip.local/oauth/token
          scopes:
            cases:read: 사건 조회 권한
            cases:write: 사건 작성/수정 권한
            evidences:upload: 증거 업로드 권한
            reports:read: 리포트 조회 권한
            reports:write: 리포트 작성/수정 권한
            payments:read: 결제 조회 권한
            payments:write: 결제 생성/변경 권한
            admin:all: 관리자 전체 권한
        authorizationCode:
          authorizationUrl: https://api.piip.local/oauth/authorize
          tokenUrl: https://api.piip.local/oauth/token
          scopes:
            cases:read: 사건 조회 권한
            cases:write: 사건 작성/수정 권한
            evidences:upload: 증거 업로드 권한
            reports:read: 리포트 조회 권한
            reports:write: 리포트 작성/수정 권한
            payments:read: 결제 조회 권한
            payments:write: 결제 생성/변경 권한
            admin:all: 관리자 전체 권한
  parameters:
...
    ReportId:
      in: path
      name: reportId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: 유효성 검사 실패
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Error"
              - type: object
                properties:
                  details:
                    type: object
                    additionalProperties: true
  schemas:
...
    Report:
      type: object
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        status: { type: string, enum: [draft, in_review, approved, rejected] }
        version: { type: integer }
        title: { type: string }
        summary: { type: string }
        body: { type: object }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...